+++ source/SoloFeature_outputResults.cpp
@@ -29,16 +29,16 @@ void SoloFeature::outputResults(bool cellFilterYes, string outputPrefixMat)
             ofstream &geneStr=ofstrOpen(outputPrefixMat+pSolo.outFileNames[1],ERROR_OUT, P);
             for (uint32 ii=0; ii<Trans.nGe; ii++) {
                 geneStr << Trans.geID[ii] <<"\t"<< (Trans.geName[ii].empty() ? Trans.geID[ii] : Trans.geName[ii]);
-				if (pSolo.outFormat.featuresGeneField3!="-") {
-					geneStr <<'\t'<< pSolo.outFormat.featuresGeneField3;
-				};
-				geneStr << '\n';
+                if (pSolo.outFormat.featuresGeneField3!="-") {
+                    geneStr <<'\t'<< pSolo.outFormat.featuresGeneField3;
+                };
+                geneStr << '\n';
             };
             geneStr.close();
             break;
         };
         case SoloFeatureTypes::SJ :
-        	symlink("../../../SJ.out.tab", (outputPrefixMat+pSolo.outFileNames[1]).c_str());
+            symlink("../../../SJ.out.tab", (outputPrefixMat+pSolo.outFileNames[1]).c_str());
     };
 
     ////////////////////////////////////////////////////////////////////////////
@@ -113,62 +113,101 @@ void SoloFeature::outputResults(bool cellFilterYes, string outputPrefixMat)
     
     //////////////////////////////////////////// output unique+multimappers
     if (pSolo.multiMap.yes.multi && !cellFilterYes && (featureType == SoloFeatureTypes::Gene || featureType == SoloFeatureTypes::GeneFull) ) {
-        
+                          //skipping unique
         for (uint32 iMult=1; iMult<pSolo.multiMap.tN; iMult++) {
             if (!pSolo.multiMap.yes.B[iMult])
                 continue; //no output for this type
                 
-            
-            for (uint32 iDed=1; iDed<countMatStride; iDed++) {
+            for (uint32 iDed=0; iDed<pSolo.umiDedup.yes.N; iDed++) {
                 string matrixFileName=outputPrefixMat + "UniqueAndMult-" + pSolo.multiMap.typeNames[iMult];
-                
                 if (pSolo.umiDedup.types.size()>1) {
-                    matrixFileName += "_umiDedup-" + pSolo.umiDedup.typeNames[pSolo.umiDedup.types[iDed-1]];
+                    matrixFileName += "_umiDedup-" + pSolo.umiDedup.typeNames[pSolo.umiDedup.types[iDed]];
+                };
+                matrixFileName += ".mtx";
+
+                uint32 mIndex = 0;
+                switch(iMult) {
+                    case (pSolo.multiMap.typeI::Uniform):
+                        mIndex = 1;
+                        break;
+                    case (pSolo.multiMap.typeI::Rescue):
+                        mIndex = 2+iDed;    
                 };
                 
-                matrixFileName += ".mtx";
+                //string matOutString;
+                //matOutString.reserve(100000000);
+                std::ostringstream matOutStringStream;//(matOutString);
                 
-                string countMatOutString;
-                countMatOutString.reserve(100000000);
                 nCellGeneEntries = 0;
                 
                 uint32  cbInd1=0;
                 for (uint32 icb=0; icb<nCB; icb++) {
                     cbInd1=indCB[icb]+1;
                     
-                    /*
+                    /* just unique
                     for (uint32 igm=countCellGeneUMIindex[icb]; igm<countCellGeneUMIindex[icb+1]; igm+=countMatStride) {
                                                
                         //feature index, CB index, count
-                        countMatOutString += to_string(countCellGeneUMI[igm]+1) +' '+ to_string(cbInd1) +' '+ to_string(countCellGeneUMI[igm+iDed]) + '\n';
+                        matOutString += to_string(countCellGeneUMI[igm]+1) +' '+ to_string(cbInd1) +' '+ to_string(countCellGeneUMI[igm+iDed]) + '\n';
                         
                         ++nCellGeneEntries;
                     };
                     */
                     
-                    //sum unique+multiple
-                    for (uint32 igm1=countCellGeneUMIindex[icb]; igm1<countCellGeneUMIindex[icb+1]; igm1+=countMatStride) {
-                                               
-                        auto g1 = countCellGeneUMI[igm1];
-                        auto c1 = countCellGeneUMI[igm1+iDed];
-                        
-                        while (g2<g1 && igm2 < countMatMult.i[icb+1]) {
-                        };
-                        //feature index, CB index, count
-                        countMatOutString += to_string(g1+1) +' '+ to_string(cbInd1) +' '+ to_string(c1) +'\n';
-                        
+                    /*//just multiple
+                    for (uint32 igm=countMatMult.i[icb]; igm<countMatMult.i[icb+1]; igm+=countMatMult.s) {                      
+                        matOutStringStream << int(countMatMult.m[igm])+1 <<' '<< cbInd1 <<' '<< countMatMult.m[igm+mIndex] <<'\n';
                         ++nCellGeneEntries;
                     };
+                    */
                     
+                    //if (false)
+                    {//sum unique+multiple: go over sorted lists. TODO: use a map to combine (check efficiency)
+                        auto igm1 = countCellGeneUMIindex[icb];
+                        auto igm2 = countMatMult.i[icb];
+                        while ( igm1<countCellGeneUMIindex[icb+1] || igm2<countMatMult.i[icb+1] ) {
+                            uint32 g1,c1,g2;
+                            double c2;
+                            
+                            if (igm1<countCellGeneUMIindex[icb+1]) {
+                                g1 = countCellGeneUMI[igm1];
+                                c1 = countCellGeneUMI[igm1+1+iDed];
+                            } else {
+                                g1 = (uint32)-1;
+                            };
+                            
+                            if (igm2<countMatMult.i[icb+1]) {
+                                g2 = countMatMult.m[igm2];
+                                c2 = countMatMult.m[igm2+mIndex];                                
+                            } else {
+                                g2 = (uint32)-1;
+                            };
+                            
+                            if (g1<g2) {
+                                matOutStringStream << g1+1 <<' '<< cbInd1 <<' '<< c1 <<'\n';
+                                igm1 += countMatStride;
+                            } else if (g1>g2) {
+                                matOutStringStream << g2+1 <<' '<< cbInd1 <<' '<< c2 <<'\n';
+                                igm2 += countMatMult.s;  
+                            } else {
+                                matOutStringStream << g1+1 <<' '<< cbInd1 <<' '<< c1+c2 <<'\n';
+                                igm1 += countMatStride;
+                                igm2 += countMatMult.s;  
+                            };
+                            
+                            ++nCellGeneEntries;
+                        };
+                    };
                 };
-
-                ofstream &countMatrixStream=ofstrOpen(matrixFileName,ERROR_OUT, P);
+                //matOutStringStream.flush();
+                
+                ofstream &countMatrixStream=ofstrOpen(matrixFileName, ERROR_OUT, P);
                 
                 //header
                 countMatrixStream <<"%%MatrixMarket matrix coordinate float general\n";
                 countMatrixStream <<"%\n";
                 countMatrixStream << featuresNumber <<' '<< pSolo.cbWLsize <<' '<< nCellGeneEntries << '\n';
-                countMatrixStream << countMatOutString;
+                countMatrixStream << matOutStringStream.str();
 
                 countMatrixStream.close();
             };